Add files via upload
I use three different algorithms for VAD,then using DTW for recognition.The results of the recognition rate are shown in Table 1 and Figure 1.  Cos represents the algorithm MFCC_COS, Dist represents the original algorithm which is based on the MFCC parameters’ Euclidean distance, and Energy+Zero represents the conventional method which is based on the short-time energy and short-zero rate.